requirements.txt
# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# API & Async
httpx>=0.25.0
asyncio>=3.4.3

# File handling
PyPDF2>=3.0.0
python-docx>=0.8.11
extract-msg>=0.45.0
chardet>=5.2.0
openpyxl>=3.1.2

# Excel parsing
xlsxwriter>=3.1.0
xlrd>=2.0.1

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.1

# ML/Data processing (pour le parseur Excel)
joblib>=1.3.0
tqdm>=4.66.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0